rename dashboard filenames to be less confusing
