+++ schedview/app/scheduler_dashboard/__init__.py
@@ -2,4 +2,4 @@ __all__ = [
     "scheduler_app",
 ]
 
-from .scheduler_dashboard import scheduler_app
+from .scheduler_dashboard_app import scheduler_app
